server/requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-multipart
python-dotenv